+++ src/neptune/cli/containers.py
@@ -67,7 +67,13 @@ retries_timeout = int(os.getenv(NEPTUNE_SYNC_BATCH_TIMEOUT_ENV, "3600"))
 
 
 class ExecutionDirectory:
-    def __init__(self, path: Path, synced: bool, structure_version: StructureVersion, parent: Optional[Path] = None):
+    def __init__(
+        self,
+        path: Path,
+        synced: bool,
+        structure_version: StructureVersion,
+        parent: Optional[Path] = None,
+    ):
         self._path = path
         self._synced = synced
         self._structure_version = structure_version
@@ -93,7 +99,14 @@ class ExecutionDirectory:
         if self.path.exists():
             remove_directory_structure(self.path)
 
-    def sync(self, *, backend: "NeptuneBackend", container_id: UniqueId, container_type: ContainerType) -> None:
+    def sync(
+        self,
+        *,
+        backend: "NeptuneBackend",
+        container_id: UniqueId,
+        container_type: ContainerType,
+        no_cleanup: Optional[bool] = False,
+    ) -> None:
         operation_storage = OperationStorage(self.path)
         serializer: Callable[[Operation], Dict[str, Any]] = lambda op: op.to_dict()
 
@@ -104,6 +117,7 @@ class ExecutionDirectory:
             to_dict=serializer,
             from_dict=Operation.from_dict,
             lock=threading.RLock(),
+            no_cleanup=no_cleanup,
         ) as disk_queue:
             while True:
                 raw_batch = disk_queue.get_batch(batch_size)
@@ -138,17 +152,25 @@ class ExecutionDirectory:
                             ex.cause.__class__.__name__,
                         )
 
-    def move(self, *, base_path: Path, target_container_id: UniqueId, container_type: ContainerType) -> None:
+    def move(
+        self,
+        *,
+        base_path: Path,
+        target_container_id: UniqueId,
+        container_type: ContainerType,
+    ) -> Optional[Path]:
         new_online_dir = get_container_dir(container_id=target_container_id, container_type=container_type)
         try:
             (base_path / ASYNC_DIRECTORY).mkdir(parents=True, exist_ok=True)
         except OSError:
             logger.warning(f"Cannot create directory: {base_path / ASYNC_DIRECTORY}")
-            return
+            return None
 
-        self.path.rename(base_path / ASYNC_DIRECTORY / new_online_dir)
-        self._path = base_path / ASYNC_DIRECTORY / new_online_dir
+        new_path = base_path / ASYNC_DIRECTORY / new_online_dir
+        self.path.rename(new_path)
+        self._path = new_path
         self._structure_version = StructureVersion.DIRECT_DIRECTORY
+        return new_path
 
 
 class Container(ABC):
@@ -165,7 +187,13 @@ class Container(ABC):
         return all(map(lambda execution_dir: execution_dir.synced, self.execution_dirs))
 
     @abstractmethod
-    def sync(self, *, base_path: Path, backend: "NeptuneBackend", project: Optional["Project"] = None) -> None: ...
+    def sync(
+        self,
+        *,
+        base_path: Path,
+        backend: "NeptuneBackend",
+        project: Optional["Project"] = None,
+    ) -> None: ...
 
     def clear(self) -> None:
         for execution_dir in self.execution_dirs:
@@ -212,7 +240,14 @@ class AsyncContainer(Container):
     def experiment(self) -> Optional["ApiExperiment"]:
         return self._experiment
 
-    def sync(self, *, base_path: Path, backend: "NeptuneBackend", project: Optional["Project"] = None) -> None:
+    def sync(
+        self,
+        *,
+        base_path: Path,
+        backend: "NeptuneBackend",
+        project: Optional["Project"] = None,
+        no_cleanup: Optional[bool] = False,
+    ) -> None:
         assert self.experiment is not None  # mypy fix
 
         qualified_container_name = get_qualified_name(self.experiment)
@@ -224,10 +259,16 @@ class AsyncContainer(Container):
                     backend=backend,
                     container_id=self.container_id,
                     container_type=self.container_type,
+                    no_cleanup=no_cleanup,
                 )
 
-        self.clear()
-        logger.info("Synchronization of %s %s completed.", self.experiment.type.value, qualified_container_name)
+        if not no_cleanup:
+            self.clear()
+        logger.info(
+            "Synchronization of %s %s completed.",
+            self.experiment.type.value,
+            qualified_container_name,
+        )
 
 
 class OfflineContainer(Container):
@@ -259,7 +300,14 @@ class OfflineContainer(Container):
     def found(self) -> bool:
         return self._found
 
-    def sync(self, *, base_path: Path, backend: "NeptuneBackend", project: Optional["Project"] = None) -> None:
+    def sync(
+        self,
+        *,
+        base_path: Path,
+        backend: "NeptuneBackend",
+        project: Optional["Project"] = None,
+        no_cleanup: Optional[bool] = False,
+    ) -> None:
         assert project is not None  # mypy fix
 
         experiment = register_offline_container(
@@ -271,17 +319,26 @@ class OfflineContainer(Container):
         if experiment:
             self._container_id = experiment.id
             for execution_dir in self.execution_dirs:
-                execution_dir.move(
+                self.path = execution_dir.move(
                     base_path=base_path,
                     target_container_id=self.container_id,
                     container_type=self.container_type,
+                )
+                logger.info(
+                    "Offline container %s moved to %s",
+                    f"{self.container_type}__{self.container_id}",
+                    self.path,
                 )
         else:
             logger.warning("Cannot register offline container %s", self.container_id)
             return
 
         qualified_container_name = get_qualified_name(experiment)
-        logger.info("Offline container %s registered as %s", self.container_id, qualified_container_name)
+        logger.info(
+            "Offline container %s registered as %s",
+            self.container_id,
+            qualified_container_name,
+        )
         logger.info("Synchronising %s", qualified_container_name)
 
         for execution_dir in self.execution_dirs:
@@ -289,10 +346,17 @@ class OfflineContainer(Container):
                 backend=backend,
                 container_id=self.container_id,
                 container_type=self.container_type,
+                no_cleanup=no_cleanup,
             )
 
-        self.clear()
-        logger.info("Synchronization of %s %s completed.", experiment.type.value, qualified_container_name)
+        if not no_cleanup:
+            self.clear()
+
+        logger.info(
+            "Synchronization of %s %s completed.",
+            experiment.type.value,
+            qualified_container_name,
+        )
 
 
 def remove_directory(path: Path) -> None: